[Add/Fix] Debut de mise en place de spring security et correction de bug dans le module console

+++ console/src/main/java/com/excilys/cdb/ui/Main.java
@@ -2,6 +2,8 @@ package main.java.com.excilys.cdb.ui;
 
 import java.util.Scanner;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
@@ -10,6 +12,8 @@ import main.java.com.excilys.cdb.service.ComputerService;
 
 public class Main {
 
+    static Logger LOGGER = LoggerFactory.getLogger(Main.class);
+	
 	static CompanyService companyservice;
 	static ComputerService computerservice;
 
@@ -24,21 +28,29 @@ public class Main {
 	}
 
 	public static void main(String[] args) {
-        
+		
 		context = new ClassPathXmlApplicationContext("applicationConfig.xml");
 		computerservice = (ComputerService) context.getBean("computerservice");
 		companyservice = (CompanyService) context.getBean("companyservice");
 		
-    	new UIController(companyservice, computerservice, new Interface(new Scanner(System.in)), new Scanner(System.in)).launchUI();
-    	/*
-        System.out.print("Quel company voulez-vous supprimer ? : ");
-           
-    	try (Scanner sc = new Scanner(System.in);) {
-    		int idCompany = Integer.parseInt(sc.nextLine());
-    		companyservice.remove(idCompany);
+		System.out.print("voulez-vous supprimer une compagnie ?");
+		System.out.print("1 - Oui");
+		System.out.print("2 - Non");
+
+		Scanner sc = new Scanner(System.in);
+		
+		try {
+    		int choix = Integer.parseInt(sc.nextLine());
+    		if (choix == 1) {
+    	        System.out.print("Quel company voulez-vous supprimer ? : ");
+        		int idCompany = Integer.parseInt(sc.nextLine());
+        		companyservice.remove(idCompany);
+    		}
     	} catch (NumberFormatException e) {
-            Logger logger = LoggerFactory.getLogger("Main");
-            logger.debug("Probleme de conversion en entier.");
-    	}*/
+    		LOGGER.debug("Probleme de conversion en entier.");
+    	}
+		
+    	new UIController(companyservice, computerservice, new Interface(sc), new Scanner(System.in)).launchUI();
+    	
     }
 }
